# api.yml
#Test with: http://editor.swagger.io/

swagger: '2.0'
info:
  version: 1.0.0
  title: Microservice for Chat
  description: Js + FireStore
  license:
    name: Software Architecture, 2019-II
host: localhost:4002
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /messages:
    post:
      summary: Creates a new message
      operationId: createMessage
      parameters:
        - name: message
          in: body
          required: true
          description: Message Object
          schema:
            $ref: '#/definitions/Message'
      responses:
        '200':
          description: Send
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /messages/conversation:
    delete:
      summary: Delete a conversation
      operationId: deleteConversation
      parameters:
        - name: conversation
          in: body
          description: The conversation emails to be deleted
          required: true
          schema:
            $ref: '#/definitions/Conversation'
      responses:
        '200':
          description: Conversation deleted
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /messages/notifications:
    delete:
      summary: Delete a notifications
      operationId: deleteNotifications
      parameters:
        - name: notification
          in: body
          description: The message notifications to be deleted
          required: true
          schema:
            $ref: '#/definitions/Notification'
      responses:
        '200':
          description: Notification deleted
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

definitions:
  # Messages
  Message:
    type: object
    properties:
      message:
        type: object
        required:
          - sent
          - received
          - content
          - adopt
        properties:
          sent:
            type: string
          received:
            type: string
          content:
            type: string
          adopt:
            type: boolean

  # Conversation
  Conversation:
    type: object
    properties:
      conversation:
        type: object
        required:
          - sent
          - received
        properties:
          sent:
            type: string
          received:
            type: string
  
  # Notification
  Notification:
    type: object
    properties:
      notification:
        type: object
        required:
          - sent
          - received
        properties:
          sent:
            type: string
          received:
            type: string

  # Errors
  ErrorModel:
    type: object
    required:
      - data
    properties:
      data:
        type: object
        properties:
          errors: 
           type: string